DatabaseHandler: Fix empty check
In `transaction_metervalues_insert ()`, we were accessing the
first element of a list, then checking whether or not it was
empty. This causes an exception since if this list *is* empty,
there's no element to access.

This commit shifts the code around so that the empty check occurs
before accessing any elements of the list.

Co-authored-by: Gianfranco Berardi <54074967+gberardi-pillar@users.noreply.github.com>
Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>