ocpp::v201::utils::sha256(): drop use of deprecated OpenSSL functions
The set of SHA256_* types and functions are deprecated since OpenSSL 3.
Compilation warns
.../libocpp/lib/ocpp/v201/utils.cpp:151:16: warning: ‘int SHA256_Init(SHA256_CTX*)’ is deprecated: Since OpenSSL 3.0 [-Wdeprecated-declarations]
and may even fail if OpenSSL is built without support for the deprecated
functions.

Switch to the recommended set of EVP_* functions. These have been available
since about OpenSSL 0.9.7.

Also move the required includes out of the header file to the
implementation.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>