bug: correct precision of meter_start
The meter start value (the energy in Wh that a powermeter reports
at the beginning of a transaction) is internally represented in the
transaction class (for ocpp 1.6) as a `StampedEnergyWh`.
`StampedEnergyWh` records this value as a `double` and hence its
constructor should also accept the energy as a `double` (it used to take
an `int32`).
Furthermore, the set of `on_transaction_started` for ocpp 1.6 also only
accepted `int32` values.  These have been changed to `double` too, as
there should be no reason to truncate them before.  Furthermore the
`meter_start` argument is passed by value now.  Passing by reference for
these *simple* and small types is just not necessary.

Note: the same could be done for the `on_transaction_stopped` functions.

Signed-off-by: aw <aw@pionix.de>