fix: database creation/migration scripts updated to support      migration from earlier database versions (OCPP1.6)
Note: PRAGMA foreign_keys = ON; removed
The sqlite documentation states that it isn't used when in a
transaction: https://sqlite.org/pragma.html#pragma_foreign_keys
"This pragma is a no-op within a transaction; foreign key constraint
 enforcement may only be enabled or disabled when there is no
 pending BEGIN or SAVEPOINT."

Database migration/creation is done within a BEGIN TRANSACTION.

PRAGMA foreign_keys = ON was removed because it gives the
misleading impression that the feature is permanently enabled.
Sqlite further states that PRAGMA foreign_keys = ON has to be
enabled per database connection.

Signed-off-by: James Chapman <james.chapman@pionix.de>