fix(generator): Fix generator for required arrays of string enums
Prior to this commit we could not handle arrays of string enums as we
expected integers since we used the default assignment. This commit
preserves the current implementation (not forcing array of objects to be
constructed by "push back") and only force arrays of enums to be done
more tediously.

Signed-off-by: Martin Litre <mnlitre@gmail.com>